# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskFlow API
  version: 1.0.0
  description: Task management backend with teams, projects, RBAC, notifications, and full-text search
servers:
  - url: http://localhost:3000/api/v1
tags:
  - name: Auth
  - name: Users
  - name: Teams
  - name: Projects
  - name: Tasks
  - name: Notifications
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                name: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /auth/verify-email:
    get:
      tags: [Auth]
      summary: Verify email and activate account
      parameters:
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Tokens issued
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: OK
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout and revoke refresh token
      responses:
        '200':
          description: OK
  /auth/change-password:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Change password
      responses:
        '200':
          description: OK
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      responses:
        '200':
          description: OK
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      responses:
        '200':
          description: OK
  /users/me:
    get:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: Get current user profile
      responses:
        '200':
          description: OK
    patch:
      tags: [Users]
      security: [{ bearerAuth: [] }]
      summary: Update profile
      responses:
        '200':
          description: OK
  /teams:
    get:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: List my teams
      responses:
        '200':
          description: OK
    post:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: Create team
      responses:
        '201':
          description: Created
  /teams/{id}:
    get:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
    patch:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: Update team (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
  /teams/{id}/invite:
    post:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: Invite user by email
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '201':
          description: Invitation sent
  /teams/invite/{token}/accept:
    post:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: Accept team invitation
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
  /teams/{teamId}/members/{userId}:
    patch:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: Update member role
      responses:
        '200':
          description: OK
    delete:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: Remove team member
      responses:
        '200':
          description: OK
  /teams/{teamId}/leave:
    delete:
      tags: [Teams]
      security: [{ bearerAuth: [] }]
      summary: Leave team
      responses:
        '200':
          description: OK
  /projects/teams/{teamId}:
    get:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      summary: List team projects
      responses:
        '200':
          description: OK
    post:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      summary: Create project
      responses:
        '201':
          description: Created
  /projects/{projectId}:
    get:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
    patch:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
    delete:
      tags: [Projects]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
  /tasks/search:
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: Full-text search tasks (PostgreSQL GIN)
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: projectId
          in: query
          schema: { type: string, format: uuid }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Paginated search results
  /tasks/projects/{projectId}:
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: List project tasks
      responses:
        '200':
          description: OK
    post:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      summary: Create task
      responses:
        '201':
          description: Created
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
    patch:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
    delete:
      tags: [Tasks]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
  /notifications:
    get:
      tags: [Notifications]
      security: [{ bearerAuth: [] }]
      summary: List notifications
      parameters:
        - name: unread
          in: query
          schema: { type: boolean }
      responses:
        '200':
          description: OK